Get both bundle and app server logging working, oracleServer does not compile
