+++ app/bundle/src/main/scala/org/bitcoins/bundle/util/BitcoinSAppJFX3.scala
@@ -1,0 +1,6 @@
+package org.bitcoins.bundle.util
+
+import org.bitcoins.server.util.BitcoinSApp
+import scalafx.application.JFXApp3
+
+trait BitcoinSAppJFX3 extends BitcoinSApp with JFXApp3
