+++ app/oracle-server/src/main/scala/org/bitcoins/oracle/server/OracleServerMain.scala
@@ -1,22 +1,21 @@
 package org.bitcoins.oracle.server
 
 import akka.actor.ActorSystem
+import org.bitcoins.db.util.DatadirParser
 import org.bitcoins.dlc.oracle.config.DLCOracleAppConfig
 import org.bitcoins.server.routes.{BitcoinSRunner, Server}
-import org.bitcoins.server.util.BitcoinSApp
+import org.bitcoins.server.util.{BitcoinSApp, BitcoinSAppScalaDaemon}
 
 import scala.concurrent.Future
 
 class OracleServerMain(override val args: Array[String])(implicit
-    override val system: ActorSystem)
+    override val system: ActorSystem,
+    conf: DLCOracleAppConfig)
     extends BitcoinSRunner {
-
-  implicit val conf: DLCOracleAppConfig =
-    DLCOracleAppConfig(datadir, baseConfig)
 
   override def start(): Future[Unit] = {
 
-    val bindConfOpt = rpcBindOpt match {
+    val bindConfOpt = conf.rpcBindOpt match {
       case Some(rpcbind) => Some(rpcbind)
       case None          => conf.rpcBindOpt
     }
@@ -59,9 +58,22 @@ class OracleServerMain(override val args: Array[String])(implicit
   }
 }
 
-object OracleServerMain extends BitcoinSApp {
+object OracleServerMain extends BitcoinSAppScalaDaemon {
 
   override val actorSystemName =
     s"bitcoin-s-oracle-${System.currentTimeMillis()}"
-  new OracleServerMain(args).run(Some("oracle"))
+
+  /** Directory specific for current network or custom dir */
+  override val customFinalDirOpt: Option[String] = Some("oracle")
+
+  val datadirParser =
+    DatadirParser(args.toVector, networkOpt, customFinalDirOpt)
+
+  System.setProperty("bitcoins.log.location", datadirParser.usedDir.toString)
+
+  implicit lazy val conf: DLCOracleAppConfig =
+    DLCOracleAppConfig(datadirParser.datadir, datadirParser.baseConfig)(
+      system.dispatcher)
+
+  new OracleServerMain(args).run()
 }
